+++ src/structures/Message.ts
@@ -1,4 +1,4 @@
-import type { APIChatMessage, APIGroupMessage } from 'groupme-api-types'
+import type { APIChatMessage, APIGroupMessage, DeleteGroupMessageResponse } from 'groupme-api-types'
 import type { Attachment, Channel, Client } from '..'
 import { Base, User } from '..'
 
@@ -7,7 +7,7 @@ interface MessageInterface<T extends Channel> {
     reply(message: string): Promise<Message<T>>
     like(): Promise<this>
     unlike(): Promise<this>
-    delete(): Promise<this>
+    delete(): Promise<void>
     get canDelete(): boolean
 }
 
@@ -68,8 +68,11 @@ export default abstract class Message<T extends Channel> extends Base implements
     unlike(): Promise<this> {
         throw new Error('Method not implemented.')
     }
-    delete(): Promise<this> {
-        throw new Error('Method not implemented.')
+    async delete(): Promise<void> {
+        await this.user.client.rest.api<DeleteGroupMessageResponse>(
+            'DELETE',
+            `conversations/${this.channel.id}/messages/${this.id}`,
+        )
     }
     get canDelete(): boolean {
         throw new Error('Method not implemented.')